+++ lib/ruby-lokalise-api.rb
@@ -29,6 +29,7 @@ require 'ruby-lokalise-api/resources/team'
 require 'ruby-lokalise-api/resources/order'
 require 'ruby-lokalise-api/resources/payment_card'
 require 'ruby-lokalise-api/resources/translation_provider'
+require 'ruby-lokalise-api/resources/team_user_group'
 
 require 'ruby-lokalise-api/collections/base'
 require 'ruby-lokalise-api/collections/project'
@@ -48,6 +49,7 @@ require 'ruby-lokalise-api/collections/screenshot'
 require 'ruby-lokalise-api/collections/order'
 require 'ruby-lokalise-api/collections/payment_card'
 require 'ruby-lokalise-api/collections/translation_provider'
+require 'ruby-lokalise-api/collections/team_user_group'
 
 require 'ruby-lokalise-api/client'