+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 2.2.0 (19-May-19)
+
+* Added support for [`TeamUserGroup` endpoint](https://lokalise.co/api2docs/ruby/#resource-team-user-groups)
+* Enhancements to method chaining
+
 ## 2.1.1 (17-May-19)
 
 * Incorporate API updates (new attributes, mostly creation and update timestamps)